Simplify coco-accelerated implementation.
