--- coco-accelerated/src/eval/futhark_cuda.rs
@@ -1,5 +1,0 @@
-use crate::{declare_eval, declare_params, declare_problem};
-
-declare_params!(coco_futhark::backend::Cuda);
-declare_eval!(coco_futhark::backend::Cuda);
-declare_problem!(coco_futhark::backend::Cuda);
