--- coco-accelerated/src/eval/mod.rs
@@ -1,15 +1,0 @@
-#![allow(dead_code)]
-
-mod futhark;
-
-#[cfg(feature = "c")]
-pub mod futhark_c;
-
-#[cfg(feature = "multicore")]
-pub mod futhark_multicore;
-
-#[cfg(feature = "opencl")]
-pub mod futhark_opencl;
-
-#[cfg(feature = "cuda")]
-pub mod futhark_cuda;
